wip on adding stop stream button
