+++ llm_studio/app_utils/sections/chat.py
@@ -1,9 +1,6 @@
-import asyncio
 import gc
 import logging
 import os
-import threading
-from typing import Any, Callable, Dict, List, Optional
 
 import numpy as np
 import torch
@@ -12,8 +9,8 @@ from accelerate.utils import get_balanced_memory
 from h2o_wave import Q
 from h2o_wave import data as chat_data
 from h2o_wave import ui
-from transformers import AutoTokenizer, TextStreamer
 
+from llm_studio.app_utils.sections.chat_update import USER, BOT, predict_text
 from llm_studio.app_utils.utils import (
     get_experiments,
     get_ui_elements,
@@ -30,54 +27,6 @@ from llm_studio.src.utils.config_utils import (
 from llm_studio.src.utils.modeling_utils import load_checkpoint
 
 logger = logging.getLogger(__name__)
-
-USER = True
-BOT = False
-
-
-class WaveChatStreamer(TextStreamer):
-    def __init__(
-        self,
-        tokenizer: AutoTokenizer,
-        q: Q,
-        text_cleaner: Optional[Callable] = None,
-        **decode_kwargs,
-    ):
-        """
-        Updates the chabot card in a streaming fashion
-        """
-        super().__init__(tokenizer, skip_prompt=True, **decode_kwargs)
-        self.text_cleaner = text_cleaner
-        self.words_predicted_answer: List[str] = []
-        self.q = q
-        self.loop = asyncio.get_event_loop()
-        self.finished = False
-
-    def on_finalized_text(self, text: str, stream_end: bool = False):
-        self.words_predicted_answer += [text]
-        self.loop.create_task(self.push_to_chat())
-
-    async def push_to_chat(self):
-        self.q.page["experiment/display/chat"].data[-1] = [self.answer, BOT]
-        await self.q.page.save()
-
-    @property
-    def answer(self):
-        """
-        Create the answer by joining the generated words.
-        By this, self.text_cleaner does not need to be idempotent.
-        """
-        answer = "".join(self.words_predicted_answer)
-        if answer.endswith(self.tokenizer.eos_token):
-            # text generation is stopped
-            answer = answer.replace(self.tokenizer.eos_token, "")
-        if self.text_cleaner:
-            answer = self.text_cleaner(answer)
-        return answer
-
-    def end(self):
-        super().end()
-        self.finished = True
 
 
 async def chat_tab(q: Q, load_model=True):
@@ -148,11 +97,17 @@ async def chat_tab(q: Q, load_model=True):
     q.page["experiment/display/chat/settings"] = ui.form_card(
         box="second",
         items=[
+            ui.buttons([
             ui.button(
                 name="experiment/display/chat/clear_history",
                 label="Clear History",
                 primary=True,
             ),
+            ui.button(
+                name="experiment/display/chat/abort_stream",
+                label="Stop Straming",
+                primary=True,
+            )]),
             ui.expander(
                 name="chat_settings",
                 label="Chat Settings",
@@ -222,7 +177,13 @@ async def chat_update(q: Q) -> None:
     Update the chatbot with the new message.
     """
     q.client["experiment/display/chat/finished"] = False
+    try:
+        await update_chat_window(q)
+    finally:
+        q.client["experiment/display/chat/finished"] = True
 
+
+async def update_chat_window(q):
     cfg_prediction = parse_ui_elements(
         cfg=q.client["experiment/display/chat/cfg"].prediction,
         q=q,
@@ -230,19 +191,15 @@ async def chat_update(q: Q) -> None:
     )
     logger.info(f"Using chatbot config: {cfg_prediction}")
     q.client["experiment/display/chat/cfg"].prediction = cfg_prediction
-
     prompt = q.client["experiment/display/chat/chatbot"]
-
     message = [prompt, USER]
     q.client["experiment/display/chat/messages"].append(message)
     q.page["experiment/display/chat"].data += message
     q.page["experiment/display/chat"].data += ["", BOT]
     await q.page.save()
-
     cfg = q.client["experiment/display/chat/cfg"]
     model: Model = q.client["experiment/display/chat/model"]
     tokenizer = q.client["experiment/display/chat/tokenizer"]
-
     full_prompt = ""
     if len(q.client["experiment/display/chat/messages"]):
         for prev_message in q.client["experiment/display/chat/messages"][
@@ -274,50 +231,14 @@ async def chat_update(q: Q) -> None:
             output={"predicted_text": np.array([text])}, cfg=cfg
         )["predicted_text"][0]
 
-    if cfg.prediction.num_beams == 1:
-        streamer = WaveChatStreamer(tokenizer=tokenizer, q=q, text_cleaner=text_cleaner)
-        # Need to start generation in a separate thread, otherwise streaming is blocked
-        thread = threading.Thread(
-            target=generate,
-            kwargs=dict(model=model, inputs=inputs, cfg=cfg, streamer=streamer),
-        )
-        try:
-            thread.start()
-        finally:
-            while True:
-                if streamer.finished:
-                    thread.join()
-                    predicted_text = streamer.answer
-                    break
-                await q.sleep(1)
-    else:
-        # ValueError: `streamer` cannot be used with beam search (yet!).
-        # Make sure that `num_beams` is set to 1.
-        logger.info("Not streaming output, as it cannot be used with beam search.")
-        q.page["experiment/display/chat"].data[-1] = ["...", BOT]
-        await q.page.save()
-        predicted_answer_ids = generate(model, inputs, cfg)[0]
-        predicted_text = tokenizer.decode(
-            predicted_answer_ids, skip_special_tokens=True
-        )
-        predicted_text = text_cleaner(predicted_text)
-
+    predicted_text = await predict_text(cfg, inputs, model, q, text_cleaner, tokenizer)
     logger.info(f"Predicted Answer: {predicted_text}")
     message = [predicted_text, BOT]
     q.client["experiment/display/chat/messages"].append(message)
     q.page["experiment/display/chat"].data[-1] = message
-
     del inputs
     gc.collect()
     torch.cuda.empty_cache()
-
-    q.client["experiment/display/chat/finished"] = True
-
-
-def generate(model: Model, inputs: Dict, cfg: Any, streamer: TextStreamer = None):
-    with torch.cuda.amp.autocast():
-        output = model.generate(batch=inputs, cfg=cfg, streamer=streamer).detach().cpu()
-    return output
 
 
 def load_cfg_model_tokenizer(
@@ -383,3 +304,14 @@ async def show_chat_is_running_dialog(q):
         closable=True,
     )
     await q.page.save()
+
+
+async def block_app_while_streaming(q: Q):
+    if q.args["experiment/display/chat/abort_stream"]:
+        # - join thread and check it has been finished
+        # - set "experiment/display/chat/finished" to False
+        return True
+    elif q.client["experiment/display/chat/finished"] is False:
+        await show_chat_is_running_dialog(q)
+        return True
+    return False
