+++ model_cards/text_causal_language_modeling_model_card_template.md
@@ -105,6 +105,7 @@ print(res[0]["generated_text"])
 You may also construct the pipeline from the loaded model and tokenizer yourself and consider the preprocessing steps:
 
 ```python
+import llm_studio.app_utils.sections.chat_update
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 model_name = "{{repo_id}}"  # either local folder or huggingface model name
@@ -127,7 +128,7 @@ model.cuda().eval()
 inputs = tokenizer(prompt, return_tensors="pt", add_special_tokens=False).to("cuda")
 
 # generate configuration can be modified to your needs
-tokens = model.generate(
+tokens = llm_studio.app_utils.sections.chat_update.generate(
     input_ids=inputs["input_ids"],
     attention_mask=inputs["attention_mask"],
     min_new_tokens={{min_new_tokens}},
